Linux build support (integrate SlickAmogus SH PR#22 + PsyCross PR#3), Windows-safe
Adds Linux/POSIX build support without breaking the Windows (MinGW) build. Taken
from the Linux PR (preferred over the macOS PR's per-site casts because the
structural fix subsumes them):
 - include/decomp/types.h: VECTOR3 long->s32 (LP64 fix; identical size on Windows)
 - pc_port/include/psyq/libsnd.h: VabHdr/ProgAtr long->int (disc-overlaid layout)
 - pc_port/src/main_pc.c: SH_NULL_DEVICE + #ifdef _WIN32 AllocConsole
 - pc_port/src/pc_crash.c: #ifdef _WIN32 SEH path + POSIX no-op
 - pc_port/build.sh: platform exe name
 - pc_port/CMakeLists.txt: CMAKE_POSITION_INDEPENDENT_CODE + GCC14 -Wno-error block
 - pc_port/maps/CMakeLists.txt: if(WIN32)/else() link split (.so on ELF) + GCC14 block
 - PsyCross submodule -> 2338141 (= SlickAmogus/PsyCross PR#3 Linux commit 0c0b9a2,
   with our parked PPF-1 scaffold on top): portable OpenAL find_package, hidden
   visibility for libapi.c/PsyX_SPUAL.cpp, extern strcasecmp decl.

IMPORTANT correctness fix vs the PR as written: two of its -Wno-error= flags
(-Wdeclaration-missing-parameter-type, -Wreturn-mismatch) are GCC-14-only, and GCC
errors hard on -Wno-error=<unknown-flag>, which broke our GCC-13 Windows build.
Gated those blocks (exe + maps) on GCC>=14 so they're a no-op on Windows and active
on the Linux GCC-14 target. Windows build verified: exe + 42/42 map DLLs, 0 errors.
macOS (arm64) build infra is APPLE-gated and can layer on later. PsyCross commit
2338141 is local — push it to SlickAmogus/PsyCross before publishing the gitlink.

+++ pc_port/include/psyq/libsnd.h
@@ -59,11 +59,18 @@
 #endif
 
 /* VAB structures */
+/* NOTE: these structs are overlaid directly on the raw VAB data read from
+ * disc, so their layout MUST match the PSX's. PSX `long` is 32-bit; using the
+ * C `long` here is correct on Windows (LLP64, long==4) but WRONG on Linux/
+ * macOS (LP64, long==8), where it shifts every following field and inflates
+ * sizeof — fsize/ps were read at the wrong offset, yielding a garbage VAB
+ * size and an SPU-malloc failure that crashed VAB transfer. Use fixed-width
+ * 32-bit ints so the layout is correct on every target. */
 typedef struct VabHdr {
-    long           form;
-    long           ver;
-    long           id;
-    unsigned long  fsize;
+    int            form;
+    int            ver;
+    int            id;
+    unsigned int   fsize;
     unsigned short reserved0;
     unsigned short ps;
     unsigned short ts;
@@ -72,7 +79,7 @@ typedef struct VabHdr {
     unsigned char  pan;
     unsigned char  attr1;
     unsigned char  attr2;
-    unsigned long  reserved1;
+    unsigned int   reserved1;
 } VabHdr;
 
 typedef struct ProgAtr {
@@ -83,8 +90,8 @@ typedef struct ProgAtr {
     unsigned char mpan;
     char          reserved0;
     short         attr;
-    unsigned long reserved1;
-    unsigned long reserved2;
+    unsigned int  reserved1;
+    unsigned int  reserved2;
 } ProgAtr;
 
 typedef struct VagAtr {
